@@ -10,9 +10,11 @@
     "@types/node": "^16.11.17",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-rc.0",
+    "react-dom": "^18.0.0-rc.0",
     "react-scripts": "5.0.0",
+    "sleep-promise": "^9.1.0",
+    "suspensify": "^1.0.1",
     "typescript": "^4.5.4",
     "web-vitals": "^2.1.2"
   },
