@@ -1,26 +1,25 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import logo from "./logo.svg";
+import "./App.css";
+import ErrorBoundary from "./error-boundary";
+import { Suspense } from "react";
+import PokemonDetail from "./pokemon-detail";
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
-}
+const App = () => {
+    return (
+        <div className="App">
+            <h1>Pokedex memory game</h1>
+            <p>
+                Test your Poke-memory by guessing which pokemon is next - before
+                it appears.
+            </p>
+
+            <ErrorBoundary>
+                <Suspense fallback="loading pokemon...">
+                    <PokemonDetail />
+                </Suspense>
+            </ErrorBoundary>
+        </div>
+    );
+};
 
 export default App;
